# add lz4 compression for jackson