--- a/akka-bench-jmh/src/main/scala/akka/serialization/jackson/JacksonSerializationBench.scala
+++ b/akka-bench-jmh/src/main/scala/akka/serialization/jackson/JacksonSerializationBench.scala
@@ -187,8 +187,13 @@ class JacksonSerializationBench {
   var serialization: Serialization = _
 
   @silent("immutable val") // JMH updates this via reflection
-  @Param(Array("jackson-json", "jackson-cbor")) // "java"
+  @Param(Array("jackson-json")) // "java"
   private var serializerName: String = _
+
+  @silent("immutable val")
+//  @Param(Array("off", "gzip", "lz4"))
+  @Param(Array("lz4"))
+  private var compression: String = _
 
   @Setup(Level.Trial)
   def setupTrial(): Unit = {
@@ -208,7 +213,7 @@ class JacksonSerializationBench {
           }
         }
         akka.serialization.jackson.jackson-json.compression {
-          algorithm = off
+          algorithm = $compression
           compress-larger-than = 100 b
         }
       """)
@@ -222,10 +227,18 @@ class JacksonSerializationBench {
     Await.result(system.terminate(), 5.seconds)
   }
 
+  private var size = 0L
+
   private def serializeDeserialize[T <: AnyRef](msg: T): T = {
     serialization.findSerializerFor(msg) match {
       case serializer: SerializerWithStringManifest =>
         val blob = serializer.toBinary(msg)
+        if (size != blob.length) {
+          size = blob.length
+          println(
+            s"# Size is $size of ${msg.getClass.getName} with " +
+            s"${system.settings.config.getString("akka.serialization.jackson.jackson-json.compression.algorithm")}")
+        }
         serializer.fromBinary(blob, serializer.manifest(msg)).asInstanceOf[T]
       case serializer =>
         val blob = serializer.toBinary(msg)
@@ -237,39 +250,39 @@ class JacksonSerializationBench {
 
   }
 
-  @Benchmark
-  def small(): Small = {
-    serializeDeserialize(smallMsg1)
-  }
-
-  @Benchmark
-  def medium(): Medium = {
-    serializeDeserialize(mediumMsg1)
-  }
+//  @Benchmark
+//  def small(): Small = {
+//    serializeDeserialize(smallMsg1)
+//  }
+//
+//  @Benchmark
+//  def medium(): Medium = {
+//    serializeDeserialize(mediumMsg1)
+//  }
 
   @Benchmark
   def large(): Large = {
     serializeDeserialize(largeMsg)
   }
 
-  @Benchmark
-  def jSmall(): JSmall = {
-    serializeDeserialize(jSmallMsg1)
-  }
-
-  @Benchmark
-  def jMedium(): JMedium = {
-    serializeDeserialize(jMediumMsg1)
-  }
-
-  @Benchmark
-  def jLarge(): JLarge = {
-    serializeDeserialize(jLargeMsg)
-  }
-
-  @Benchmark
-  def timeMessage(): TimeMessage = {
-    serializeDeserialize(timeMsg)
-  }
+//  @Benchmark
+//  def jSmall(): JSmall = {
+//    serializeDeserialize(jSmallMsg1)
+//  }
+//
+//  @Benchmark
+//  def jMedium(): JMedium = {
+//    serializeDeserialize(jMediumMsg1)
+//  }
+//
+//  @Benchmark
+//  def jLarge(): JLarge = {
+//    serializeDeserialize(jLargeMsg)
+//  }
+//
+//  @Benchmark
+//  def timeMessage(): TimeMessage = {
+//    serializeDeserialize(timeMsg)
+//  }
 
 }