--- a/akka-serialization-jackson/src/main/resources/reference.conf
+++ b/akka-serialization-jackson/src/main/resources/reference.conf
@@ -191,6 +191,7 @@ akka.serialization.jackson.jackson-json.compression {
   # Compression algorithm.
   # - off  : no compression
   # - gzip : using common java gzip
+  # - lz4 : using lz4-java
   algorithm = gzip
 
   # If compression is enabled with the `algorithm` setting the payload is compressed