… akka-docs/src/main/paradox/serialization-jackson.md
… akka-serialization-jackson/src/test/scala/akka/serialization/jackson/JacksonSerializerSpec.scala