References: https://github.com/akka/akka/issues/27066 https://github.com/akka/akka/pull/27131