@patriknw 2020-05-05
I tried with the JMH benchmark, and the results are important for the choice here. Pushed that to a temporary branch https://github.com/akka/akka/tree/wip-29012-lz4-bench-patriknw and you can incorporate the ideas from that here. See the [last 3 commits](https://github.com/akka/akka/commits/wip-29012-lz4-bench-patriknw) there.

**Baseline, this PR**
```
[info] Benchmark                                                         (compression)  (serializerName)   Mode  Cnt         Score        Error   Units
[info] JacksonSerializationBench.large                                             off      jackson-json  thrpt    5     17201,643 ±    280,200   ops/s
[info] JacksonSerializationBench.large:·gc.alloc.rate                              off      jackson-json  thrpt    5      1049,425 ±     18,645  MB/sec
[info] JacksonSerializationBench.large:·gc.alloc.rate.norm                         off      jackson-json  thrpt    5     67249,508 ±     10,423    B/op
[info] JacksonSerializationBench.large                                            gzip      jackson-json  thrpt    5      9524,178 ±     78,544   ops/s
[info] JacksonSerializationBench.large:·gc.alloc.rate                             gzip      jackson-json  thrpt    5       726,130 ±      4,308  MB/sec
[info] JacksonSerializationBench.large:·gc.alloc.rate.norm                        gzip      jackson-json  thrpt    5     84026,732 ±     18,901    B/op
[info] JacksonSerializationBench.large                                             lz4      jackson-json  thrpt    5       801,604 ±     68,347   ops/s
[info] JacksonSerializationBench.large:·gc.alloc.rate                              lz4      jackson-json  thrpt    5     12276,240 ±   1096,748  MB/sec
[info] JacksonSerializationBench.large:·gc.alloc.rate.norm                         lz4      jackson-json  thrpt    5  16876505,432 ±    233,390    B/op
```

Note that 801 ops/s is extremely poor.

**LZ4FrameOutputStream(bos, LZ4FrameOutputStream.BLOCKSIZE.SIZE_64KB)**

Changing the block size.

```
[info] Benchmark                                                     (compression)  (serializerName)   Mode  Cnt       Score       Error   Units
[info] JacksonSerializationBench.large                                         lz4      jackson-json  thrpt    5   12082,897 ±  1105,616   ops/s
[info] JacksonSerializationBench.large:·gc.alloc.rate                          lz4      jackson-json  thrpt    5    3784,992 ±   345,228  MB/sec
[info] JacksonSerializationBench.large:·gc.alloc.rate.norm                     lz4      jackson-json  thrpt    5  345250,149 ±    14,849    B/op
```

12082 ops/s is better, but not extremely impressive. 17201 ops/s is without compression.

**raw  LZ4Factory**
```
[info] Benchmark                                                     (compression)  (serializerName)   Mode  Cnt      Score       Error   Units
[info] JacksonSerializationBench.large                                         lz4      jackson-json  thrpt    5  15252,463 ±   550,351   ops/s
[info] JacksonSerializationBench.large:·gc.alloc.rate                          lz4      jackson-json  thrpt    5   1179,044 ±    40,124  MB/sec
[info] JacksonSerializationBench.large:·gc.alloc.rate.norm                     lz4      jackson-json  thrpt    5  85193,699 ±    11,738    B/op
```

Now we are talking. 15252 ops/s. I think there are still room for improvement when it comes to the generated garbage if we use ByteBuffer.

You can run the benchmark in that branch from sbt project akka-bench-jmh:

```
jmh:run -i 10 -wi 5 -f1 -t1 akka.actor.typed.TypedActorBenchmark
```

The drawaback of using the raw  LZ4Factory is as you say that we don't get the magic header for free and we must also include the decompressed size as meta data. However, adding those two fields in front of the payload shouldn't be too difficult.