@contrun 2020-05-07
@patriknw 

The performance now is more acceptable.

```
[info] Benchmark                              (compression)  (serializerName)   Mode  Cnt       Score       Error  Units
[info] JacksonSerializationBench.jLarge                 off      jackson-json  thrpt   10   10673,240 ±   341,314  ops/s
[info] JacksonSerializationBench.jLarge                gzip      jackson-json  thrpt   10    6820,026 ±   220,070  ops/s
[info] JacksonSerializationBench.jLarge                 lz4      jackson-json  thrpt   10   10199,358 ±   365,615  ops/s
[info] JacksonSerializationBench.jMedium                off      jackson-json  thrpt   10   70507,716 ±  3163,379  ops/s
[info] JacksonSerializationBench.jMedium               gzip      jackson-json  thrpt   10   22062,973 ±  1154,340  ops/s
[info] JacksonSerializationBench.jMedium                lz4      jackson-json  thrpt   10   61368,029 ±  2336,531  ops/s
[info] JacksonSerializationBench.jSmall                 off      jackson-json  thrpt   10  653228,830 ± 19562,622  ops/s
[info] JacksonSerializationBench.jSmall                gzip      jackson-json  thrpt   10  649666,717 ± 38353,082  ops/s
[info] JacksonSerializationBench.jSmall                 lz4      jackson-json  thrpt   10  607266,189 ± 35377,750  ops/s
[info] JacksonSerializationBench.large                  off      jackson-json  thrpt   10    9547,133 ±   347,849  ops/s
[info] JacksonSerializationBench.large                 gzip      jackson-json  thrpt   10    6291,867 ±   250,375  ops/s
[info] JacksonSerializationBench.large                  lz4      jackson-json  thrpt   10    9240,936 ±   467,253  ops/s
[info] JacksonSerializationBench.medium                 off      jackson-json  thrpt   10   62919,987 ±  6908,504  ops/s
[info] JacksonSerializationBench.medium                gzip      jackson-json  thrpt   10   20756,459 ±   607,306  ops/s
[info] JacksonSerializationBench.medium                 lz4      jackson-json  thrpt   10   58356,441 ±  2318,214  ops/s
[info] JacksonSerializationBench.small                  off      jackson-json  thrpt   10  545373,683 ± 20993,620  ops/s
[info] JacksonSerializationBench.small                 gzip      jackson-json  thrpt   10  592073,456 ± 24473,061  ops/s
[info] JacksonSerializationBench.small                  lz4      jackson-json  thrpt   10  582853,048 ± 20019,870  ops/s
[info] JacksonSerializationBench.timeMessage            off      jackson-json  thrpt   10  101474,234 ±  6022,707  ops/s
[info] JacksonSerializationBench.timeMessage           gzip      jackson-json  thrpt   10  104539,785 ±  5554,102  ops/s
[info] JacksonSerializationBench.timeMessage            lz4      jackson-json  thrpt   10  102757,500 ±  5517,842  ops/s

```

The only difference between our code is that I prepended the length of the decompressed data.
It has noticble impact on the performance as truncating array in java requires memory reallocation.
The documentation of lz4 java explicitly states this.
https://lz4.github.io/lz4-java/1.7.1/docs/net/jpountz/lz4/LZ4SafeDecompressor.html#decompress-byte:A-int-int-int-
As size of decompressed data is exact in our case. We don't have this penalty.

The compression currently requires moving data. I don't think it is possible to avoid that if we need to use the contagious bytes.
I'd like to hear your opinion.

I can allocate a larger buffer, and then write the metadata, compressed data to it. But doing so would requires truncating unused bytes, which requires memory reallocation in java.

We can delay memory reallocation by using `akka.utils.ByteString`, whose `toArray` method I believe will still reallocate memory. Should we do that?