@contrun 2020-05-07
https://jenkins.akka.io:8498/job/pr-validator-per-commit-jenkins/19794/testReport/akka.actor.typed.scaladsl/MessageAdapterSpec/log_wrapped_message_of_DeadLetter/ seems to be flaky. It succeeded on my machine.