@johanandren 2020-05-08
The message adapter failure should be fixed now #29045